Remove Rasterize, Options in Rasterizer object
