+++ raster-test/src/test/scala/geotrellis/raster/rasterize/polygon/PolygonRasterizerSpec.scala
@@ -18,7 +18,7 @@ package geotrellis.raster.rasterize.polygon
 
 import geotrellis.raster._
 import geotrellis.raster.rasterize._
-import geotrellis.raster.rasterize.Rasterize.Options
+import geotrellis.raster.rasterize.Rasterizer.Options
 import geotrellis.vector._
 import geotrellis.raster.testkit._
