+++ raster/src/main/scala/geotrellis/raster/mapalgebra/local/LocalMethods.scala
@@ -17,9 +17,6 @@
 package geotrellis.raster.mapalgebra.local
 
 import geotrellis.raster._
-import geotrellis.raster.rasterize.Rasterizer
-import geotrellis.raster.rasterize.Rasterize.Options
-import geotrellis.vector.{Geometry, Extent}
 import geotrellis.util.MethodExtensions
