+++ raster/src/main/scala/geotrellis/raster/mask/TileMaskMethods.scala
@@ -2,8 +2,8 @@ package geotrellis.raster.mask
 
 import geotrellis.raster._
 import geotrellis.raster.mapalgebra.local.{Mask, InverseMask}
-import geotrellis.raster.rasterize.Rasterize.Options
 import geotrellis.raster.rasterize.Rasterizer
+import geotrellis.raster.rasterize.Rasterizer.Options
 import geotrellis.util.MethodExtensions
 import geotrellis.vector.{Geometry, Extent}
