+++ spark/src/main/scala/geotrellis/spark/mask/Implicits.scala
@@ -1,10 +1,7 @@
 package geotrellis.spark.mask
 
-import geotrellis.raster._
-import geotrellis.raster.rasterize.Rasterize.Options
 import geotrellis.spark._
 import geotrellis.spark.TileLayerRDD
-import geotrellis.vector._
 import scala.reflect.ClassTag
