+++ spark/src/main/scala/geotrellis/spark/mask/TileLayerRDDMaskMethods.scala
@@ -1,7 +1,7 @@
 package geotrellis.spark.mask
 
 import geotrellis.raster._
-import geotrellis.raster.rasterize.Rasterize.Options
+import geotrellis.raster.rasterize.Rasterizer.Options
 import geotrellis.spark._
 import geotrellis.spark.TileLayerRDD
 import geotrellis.vector._
